Tidied up code
Split blog & blogcard
changed search select into a button
